✨ New Feature: App articles
This app can show articles of DB.

This app cannot edit, delete or save articles in database, can only read them.

This application needs to be refactored as initially it was only for items but now it manages customers and orders as well.

The only functionality you have now is to manipulate the views for the different templates.

Needs to have many more features but it is a good start.

On branch master
Changes to be committed:
	new file:   articles/__init__.py
	new file:   articles/admin.py
	new file:   articles/apps.py
	new file:   articles/migrations/0001_initial.py
	new file:   articles/migrations/__init__.py
	new file:   articles/models.py
	new file:   articles/tests.py
	new file:   articles/views.py