👌 IMPROVE: articles/views.py
I created new functions where one calls another.
Now there is a function that is in charge of returning the respective template.
This way it is more readable

The function that handles delivery will now always have the word 'deliver' at the end of it

On branch master
Changes to be committed:
	modified:   elCanario/articles/views.py