👌 IMPROVE: 🔨 Refactor Code:  views.py functions
I think it's the best refactoring I've ever done.
Now the functions are much tidier and much more readable.

In the next commits I will add documentation to understand which is the format to follow to create all the parts of the system.

On branch master
Changes to be committed:
	modified:   articles/views.py